requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
urllib3==1.26.5
Werkzeug==2.0.1
